+++ plugin_ide.core/src-lang/melnorme/lang/ide/core/SettingsChangeListener.java
@@ -1,0 +1,10 @@
+package melnorme.lang.ide.core;
+
+import melnorme.lang.ide.core.utils.prefs.IProjectPreference;
+import melnorme.lang.tooling.LocationHandle;
+
+public interface SettingsChangeListener {
+	
+	void preferenceChanged(IProjectPreference<?> setting, LocationHandle location, Object newValue);
+	
+}