+++ plugin_ide.ui/src-lang/melnorme/lang/ide/ui/editor/structure/AbstractLangStructureEditor.java
@@ -31,7 +31,6 @@ import org.eclipse.ui.views.contentoutline.IContentOutlinePage;
 
 import melnorme.lang.ide.core.LangCore;
 import melnorme.lang.ide.core.engine.IStructureModelListener;
-import melnorme.lang.ide.core.engine.LocationKey;
 import melnorme.lang.ide.core.engine.SourceModelManager;
 import melnorme.lang.ide.core.engine.SourceModelManager.StructureInfo;
 import melnorme.lang.ide.core.engine.SourceModelManager.StructureModelRegistration;
@@ -40,6 +39,7 @@ import melnorme.lang.ide.ui.editor.EditorSourceBuffer;
 import melnorme.lang.ide.ui.editor.EditorUtils;
 import melnorme.lang.ide.ui.text.AbstractLangSourceViewerConfiguration;
 import melnorme.lang.ide.ui.text.LangSourceViewerConfiguration;
+import melnorme.lang.tooling.LocationKey;
 import melnorme.lang.tooling.ast.SourceRange;
 import melnorme.lang.tooling.structure.SourceFileStructure;
 import melnorme.lang.tooling.structure.StructureElement;