+++ plugin_tooling/src/com/googlecode/goclipse/tooling/env/GoEnvironment.java
@@ -59,7 +59,7 @@ public class GoEnvironment {
 	}
 	
 	public String getGoPathString() {
-		return goPath.getGoPathWorkspaceString();
+		return goPath.getGoPathString();
 	}